% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_img.R
\docType{data}
\name{input_img}
\alias{input_img}
\title{This is data to be included as example}
\description{
This is data to be included as example
}
\author{
Xubo Yue, Chia-Wei Hsu (tester), Jian Kang (maintainer)
}
\keyword{data}
\keyword{imaging}
