% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_kCCA.R
\name{perm_kCCA}
\alias{perm_kCCA}
\title{Calculation of Strength of the Connectivity among multiple Brain Regions
This function is the core for kernel canonical correlation. Generally you do not need to use this function.}
\usage{
perm_kCCA(x, y, sig = 0.1, gama = 0.1, ncomps = 1, permNum = 500,
  kernel = "rbfdot")
}
\value{
(lists of) list of brain regions, permutation coefficient, p-value, region name and region type ("two" or "multiple").
}
\description{
Calculation of Strength of the Connectivity among multiple Brain Regions
This function is the core for kernel canonical correlation. Generally you do not need to use this function.
}
\details{
Kernel canonical correlation analysis (KCCA) can explore the nonlinear relationship between two variables.
It transformed sample vectors into the Hilbert space and maximize correlation coefficient by solving quadratically regularized Lagrangean function.
Refer to Kang's paper for more details: Kang J, Bowman FD, Mayberg H, Liu H (2016). "A depression network of functionallyconnected regions discovered via multi-attribute canonical correlation graphs."NeuroImage,141, 431-441.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/27474522}
}
\author{
Xubo Yue, Chiawei Hsu (tester), Jian Kang (maintainer)
}
