% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permkCCA_multipleRegion.R
\name{permkCCA_multipleRegion}
\alias{permkCCA_multipleRegion}
\title{Calculation of Strength of the Connectivity among multiple Brain Regions}
\usage{
permkCCA_multipleRegion(imageDat, region, resolution = "2mm",
  saveName = "None", kernel = "rbfdot", regionCode = "", niiFile2 = "",
  imgPath = getwd(), datPath = getwd(), parallel = FALSE, loc = "local",
  perm = 50, saveData = "None")
}
\arguments{
\item{imageDat}{There are two options for this argument: (1)(vectors of) imaging data (with extension .nii or .nii.gz).
You do not need to add extension in your argument, simply type in the name of file is enough. (2) the processed image data
produced by nii2RData function. No matter which option you choose, the result will be the same (as long as you use the same
dataset). When you would like to read and save nii data first and run kcca test later, you can first save output of nii2RData
and, in the future, choose option (2) to run kcca in order to save time (avoid read same dataset multiple times).}

\item{region}{user-specified multiple brain regions (as vector, for example, c(1,2,30)).}

\item{resolution}{the resolution of your region data. It can take "2mm" as default.
If user would like to use 3mm resolution, type in "3mm".}

\item{saveName}{whether to save processed imaging data.
If you do not have enough space or do not want to use space to store processed data, just type in "None" (default);
Else you need to specify name in this argument. For example, saveName="myName.RData".}

\item{kernel}{the kernel function used in training and predicting.
The default kernel is the radial basis kernel function "Gaussian" (rbfdot). Use "?kernlab::kcca" to find more available kernels.}

\item{regionCode}{the region code provided by user or default. It should have 3 columns with index, region code and region name.}

\item{niiFile2}{the nii region file you would like to use. It has default nii file and can be left as blank.}

\item{imgPath}{the directory where your nii file(s) is (are) located. It chooses your current working directory as default.}

\item{datPath}{the directory where you would like to store .RData file(s). It chooses your current working directory as default.}

\item{parallel}{whether to use parallel computing. Type FALSE as not using parallel and TRUE as using parallel. Parallel is not
recommended in local computer as it may slow down your system. Use parallel in cluster is preferred}

\item{loc}{this argument can accept argument "local" or "cluster". if you choose to use parallel computing, please specify whether
you run your code in your local computer or cluster.
If you did not choose parallel computing, then "local" or "clutser" makes no difference.}

\item{perm}{number of permutation. Default time is 50.}

\item{saveData}{whether to save output as R data. Type "None" as not save output.
Type name of file if you would like to save. For example, "output_two.RData".}
}
\value{
(lists of) list of brain regions, p-value, region type ("two" or "multiple"), and sregion name.
}
\description{
This function can calculate strength of the connectivity among multiple brain regions via kernel canonical correlation.
Permutaiton test is employed to assess the statistical significance.
}
\details{
(1) Kernel canonical correlation analysis (KCCA) can explore the nonlinear relationship between two variables.
It transformed sample vectors into the Hilbert space and maximize correlation coefficient by solving quadratically regularized
Lagrangean function.
Refer to Kang's paper for more details: Kang J, Bowman FD, Mayberg H, Liu H (2016). "A depression network of functionallyconnected
regions discovered via multi-attribute canonical correlation graphs."NeuroImage,141, 431-441.
(2) Use rgl.snapshot() function if you would like to save plot but forgot to use TRUE in screenShot argument.
}
\examples{
#result0<-permkCCA_multipleRegion(imageDat=c("preproc_con21_rest_MNI_2mm", "preproc_con23_rest_MNI_2mm"), region=c(1,60,70), regionCode="RegionList.txt", niiFile2="AAL_MNI_2mm.nii") #user can provide region data and region list
#result1<-permkCCA_multipleRegion(imageDat="UM_1_0050272_func_preproc", region=c(1,75), resolution="3mm", regionCode="RegionList90.txt", niiFile2="AAL_90_3mm.nii") #for 3mm resolution, user must provide region data and region list
#result2<-permkCCA_multipleRegion(imageDat="preproc_con21_rest_MNI_2mm", region=c(1,2,3)) #user can use default region data and region list only for 2mm resolution
#testcase1 <- nii2RData(niiFile1 = "preproc_con21_rest_MNI_2mm", saveName = "None") #one can process nii file first,
#result1<-permkCCA_multipleRegion(imageDat = testcase1, region = c(1,5,10)) #and then use processed nii file as input.
#result4<-permkCCA_multipleRegion(imageDat = testcase1, region = c(1,5)) #and then use processed nii file as input.
#result5<-permkCCA_multipleRegion(imageDat=c("preproc_con21_rest_MNI_2mm", "preproc_con23_rest_MNI_2mm"), region=c(1,60,90), regionCode="RegionList.txt", niiFile2="AAL_MNI_2mm.nii")
#summary_kcca(result1, 0.05, 1, "markdown")
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/27474522}
}
\author{
Xubo Yue, Chia-Wei Hsu (tester), Jian Kang (maintainer)
}
